# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Engineering Delivery Health Analyzer","version":"1.0.0"},"paths":{"/api/v1/issues":{"get":{"tags":["issues"],"summary":"Get Issues","operationId":"get_issues_api_v1_issues_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Issues Api V1 Issues Get"}}}}}}},"/api/v1/health-score":{"get":{"tags":["health"],"summary":"Get Health Score","operationId":"get_health_score_api_v1_health_score_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Health Score Api V1 Health Score Get"}}}}}}},"/api/v1/bottlenecks":{"get":{"tags":["bottlenecks"],"summary":"Get Bottlenecks","operationId":"get_bottlenecks_api_v1_bottlenecks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Bottlenecks Api V1 Bottlenecks Get"}}}}}}},"/api/v1/workload":{"get":{"tags":["workload"],"summary":"Get Workload","operationId":"get_workload_api_v1_workload_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Workload Api V1 Workload Get"}}}}}}},"/api/v1/jira/projects":{"get":{"tags":["jira"],"summary":"List Projects","description":"Return all JIRA projects visible to the authenticated user.","operationId":"list_projects_api_v1_jira_projects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response List Projects Api V1 Jira Projects Get"}}}}}}},"/api/v1/jira/users":{"get":{"tags":["jira"],"summary":"List Users","description":"Return assignable users for a JIRA project.","operationId":"list_users_api_v1_jira_users_get","parameters":[{"name":"project","in":"query","required":true,"schema":{"type":"string","description":"JIRA project key, e.g. TSITE","title":"Project"},"description":"JIRA project key, e.g. TSITE"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response List Users Api V1 Jira Users Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jira/sync":{"get":{"tags":["jira"],"summary":"Sync Issues","description":"Fetch issues from JIRA and write them to data/issues.json.\nClears the data_loader cache so subsequent API calls use fresh data.","operationId":"sync_issues_api_v1_jira_sync_get","parameters":[{"name":"project","in":"query","required":true,"schema":{"type":"string","description":"JIRA project key, e.g. TSITE","title":"Project"},"description":"JIRA project key, e.g. TSITE"},{"name":"assignee","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by assignee display name","title":"Assignee"},"description":"Filter by assignee display name"},{"name":"max_results","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Max issues to fetch","default":200,"title":"Max Results"},"description":"Max issues to fetch"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Sync Issues Api V1 Jira Sync Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Root  Get"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}